Builder: move node and tree to core module
The builder has two build implementations, directory, and stargz, and
we will add another build workflow named diff build next, so should
prepare some common components to the core module.
