+++ src/bin/nydus-image/core/mod.rs
@@ -2,4 +2,6 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod node;
 pub mod prefetch;
+pub mod tree;
